+++ KkthnxUI/Locales/enUS.lua
@@ -1,6 +1,6 @@
--- LOCALIZATION FOR ENUS & ENGB CLIENTS
+-- Localization FOR ENUS & ENGB CLIENTS
 
--- ANNOUNCE LOCALIZATION
+-- Announce Localization
 L_ANNOUNCE_FP_CAST = "%s is casting %s."
 L_ANNOUNCE_FP_CLICK = "%s is casting %s. Click!"
 L_ANNOUNCE_FP_PRE = "%s has prepared %s."
@@ -13,14 +13,14 @@ L_ANNOUNCE_PC_GO = "GO!"
 L_ANNOUNCE_PC_MSG = "Pulling %s in %s.."
 L_ANNOUNCE_SAPPED = "Sapped"
 L_ANNOUNCE_SAPPED_BY = "Sapped by: "
--- ARTIFACT LOCALIZATION
+-- Artifact Localization
 L_ARTIFACTBAR_CURRENTXP = "Current Experience: %s"
 L_ARTIFACTBAR_LEFTCLICK = "Left click to display the Artifact Frame"
 L_ARTIFACTBAR_TRAITS = "Traits avaiable: %s"
 L_ARTIFACTBAR_XP = "Experience: %s/%s (%d%%)"
 L_ARTIFACTBAR_XPREMAINING = "Remaining: %s"
 L_ARTIFACTBAR_XPTITLE = "Artifact Experience"
--- AUTOMATION LOCALIZATION
+-- Automation Localization
 L_SHORT_GOLD = "|cffffd700g|r"
 L_SHORT_SILVER = "|cffc7c7cfs|r"
 L_SHORT_COPPER = "|cffeda55fc|r"
@@ -31,12 +31,12 @@ L_CANT_AFFORD_REPAIR = "Repairs were unaffordable."
 L_REPAIR_NOTENOUGHMONEY = "You don't have enough money to repair!"
 L_REPAIR_REPAIRCOST = "Your items have been repaired for"
 L_MISC_SOLDTRASH = "Your vendor trash has been sold and you earned"
--- BAGS LOCALIZATION
+-- Bags Localization
 L_BAG_SHOW_BAGS = "Show Bags"
 L_BAG_RIGHT_CLICK_SEARCH = "Right-click to search"
 L_BAG_STACK_MENU = "Stack"
 L_BAG_RIGHT_CLICK_CLOSE = "Right-click to open menu"
--- BINDINGS LOCALIZATION
+-- Bindings Localization
 L_BIND_BINDING = "Binding"
 L_BIND_CLEARED = "All keybindings cleared for"
 L_BIND_DISCARD = "All newly set keybindings were discarded."
@@ -44,7 +44,7 @@ L_BIND_INSTRUCT = "Hover, your mouse over any action button, to bind it. Press t
 L_BIND_KEY = "Key"
 L_BIND_NO_SET = "No bindings set"
 L_BIND_SAVED = "All keybindings have been saved."
--- CHAT LOCALIZATION
+-- Chat Localization
 L_CHAT_AFK = "AFK"
 L_CHAT_DND = "DND"
 L_CHAT_GUILD = "G"
@@ -63,10 +63,10 @@ L_CHAT_SAYS = "says"
 L_CHAT_TRADE = "Trade"
 L_CHAT_WHISPERS = "whispers"
 L_CHAT_YELLS = "yells"
--- BIGCHAT LOCALIZATION
+-- BigChat Localization
 L_CHAT_BIGCHAT_OFF = "Big chat disabled"
 L_CHAT_BIGCHAT_ON = "Big chat enabled"
--- CombatText LOCALIZATION
+-- CombatText Localization
 L_COMBATTEXT_KILLING_BLOW = "Killing Blow"
 L_COMBATTEXT_ALREADY_UNLOCKED = "Combat text already unlocked."
 L_COMBATTEXT_ALREADY_LOCKED = "Combat text already locked."
@@ -78,7 +78,7 @@ L_COMBATTEXT_TEST_USE_TEST = "Type /xct test to toggle combat text testmode."
 L_COMBATTEXT_POPUP = "To save combat text window positions you need to reload your UI."
 L_COMBATTEXT_UNSAVED = "Combat text window positions unsaved, don't forget to reload UI."
 L_COMBATTEXT_UNLOCKED = "Combat text unlocked."
--- DATATEXT LOCALIZATION
+-- DataText Localization
 L_DATATEXT_BASESASSAULTED = "Bases Assaulted:"
 L_DATATEXT_BASESDEFENDED = "Bases Defended:"
 L_DATATEXT_CARTS_CONTROLLED = "Carts Controlled:"
@@ -92,7 +92,7 @@ L_DATATEXT_ORB_POSSESSIONS = "Orb Possessions:"
 L_DATATEXT_TOWERSASSAULTED = "Towers Assaulted:"
 L_DATATEXT_TOWERSDEFENDED = "Towers Defended:"
 L_DATATEXT_VICTORY_POINTS = "Victory Points:"
--- EXP/REP BAR LOCALIZATION
+-- Exp/Rep Bar Localization
 L_EXPERIENCE_BAR = "Experience:"
 L_EXPERIENCE_BAR_LEFTCLICK = "Left click to display the Reputation Frame"
 L_EXPERIENCE_XP = "XP: %s/%s (%d%%)"
@@ -110,12 +110,12 @@ L_REPUTATION_REP = "Rep: %s/%s (%d%%)"
 L_REPUTATION_REVERED = "Revered"
 L_REPUTATION_STANDING = "Standing: |c"
 L_REPUTATION_UNFRIENDLY = "Unfriendly"
--- IN COMBAT LOCALIZATION
+-- Incombat Localization
 L_ERR_NOT_IN_COMBAT = "You can't do that while in combat or while dead."
--- AUTOINVITE LOCALIZATION
+-- AutoInvite Localization
 L_INVITE_ENABLE = "Autoinvite enabled: "
 L_INVITE_DISABLE = "AutoInvite disabled"
--- INFO LOCALIZATION
+-- Info Localization
 L_INFO_DISBAND = "Disbanding group..."
 L_INFO_DUEL = "Declined duel request from "
 L_INFO_PET_DUEL = "Declined pet duel request from "
@@ -130,6 +130,8 @@ L_INFO_SETTINGS_ALL = "Type /settings all, to apply the settings for all modific
 L_INFO_NOT_INSTALLED = " is not installed."
 L_INFO_SKIN_DISABLED1 = "Stylization for "
 L_INFO_SKIN_DISABLED2 = " is disabled."
+-- Install Message Localization
+L_INSTALL_COMPLETE = "Installation Complete"
 -- Loot Localization
 L_LOOT_ANNOUNCE = "Announce to"
 L_LOOT_CANNOT = "Cannot roll"
@@ -143,28 +145,28 @@ L_LOOT_TO_INSTANCE = " Instance"
 L_LOOT_TO_PARTY = " Party"
 L_LOOT_TO_RAID = " Raid"
 L_LOOT_TO_SAY = " Say"
--- MAIL LOCALIZATION
+-- Mail Localization
 L_MAIL_COMPLETE = "All done."
 L_MAIL_MESSAGES = "messages"
 L_MAIL_NEED = "Need a mailbox."
 L_MAIL_STOPPED = "Stopped, inventory is full."
 L_MAIL_UNIQUE = "Stopped. Found a unique duplicate item in a bag or the bank."
--- WORLD MAP LOCALIZATION
+-- World Map Localization
 L_MAP_FOG = "Fog of War"
--- FARMMODE MINIMAP
+-- FarmMode Minimap
 L_MINIMAP_FARMMODE_ON = "Farm mode enabled"
 L_MINIMAP_FARMMODE_OFF = "Farm mode disabled"
--- MISC LOCALIZATION
+-- Misc Localization
 L_MISC_BUY_STACK = "Alt-Click to buy a stack"
 L_MISC_REPAIR = "Warning! You need to do a repair of your equipment as soon as possible!"
 L_MISC_UI_OUTDATED = "KkthnxUI is out of date. You can download the newest version from Curse.com. Get the Curse app and have KkthnxUI automatically updated with the Client!"
 L_MISC_UNDRESS = "Undress"
--- Talent spec
+-- Talent Spec Localization
 L_PLANNER_DRUID_4 = "Restoration"
 L_PLANNER_MONK_2 = "Mistweaver"
 L_PLANNER_PALADIN_1 = "Holy"
 L_PLANNER_PRIEST_1 = "Discipline"
--- POPUP LOCALIZATION
+-- Popup Localization
 L_POPUP_ARMORY = "Armory"
 L_POPUP_BOOSTUI = "|cffff0000WARNING|r |n|nThis will optimize your performance by turning down the graphics and tweaking them. Hit accept only if you are having |cffff0000FPS|r issues!|r"
 L_POPUP_DISBAND_RAID = "Are you sure you want to disband the group?"
@@ -176,7 +178,7 @@ L_POPUP_RESETUI = "Are you sure you want to reset all settings for |cff71d5ffKkt
 L_POPUP_RESOLUTIONCHANGED = "We detected a resolution change on your World of Warcraft client. We HIGHLY RECOMMEND restarting your game. Do you want to proceed?"
 L_POPUP_SETTINGS_ALL = "|cffff0000WARNING|r |n|nThis will apply all the supported addons settings and import them to go with |cff71d5ffKkthnxUI|r. This feature will not do anything if you do not have one of the supported add-ons."
 L_POPUP_SETTINGS_DBM = "We need to change the bar positions of |cff71d5ffDBM|r."
--- STATS LOCALIZATION
+-- Stats Localization
 L_STATS_GLOBAL = "Global Latency:"
 L_STATS_WORLD = "World Latency:"
 L_STATS_HOME = "Home Latency:"
@@ -186,7 +188,7 @@ L_STATS_SYSTEMLEFT = "Left Click: Open PvE-Frame"
 L_STATS_SYSTEMRIGHT = "Right Click: Clean Memory|r"
 L_STATS_DOWNLOAD = "Download: "
 L_STATS_BANDWIDTH = "Bandwidth: "
--- TOOLTIP LOCALIZATION
+-- Tooltip Localization
 L_TOOLTIP_ACH_COMPLETE = "Your Status: Completed on "
 L_TOOLTIP_ACH_INCOMPLETE = "Your Status: Incomplete"
 L_TOOLTIP_ACH_STATUS = "Your Status:"
@@ -196,9 +198,9 @@ L_TOOLTIP_ITEM_ID = "Item ID:"
 L_TOOLTIP_LOADING = "Loading..."
 L_TOOLTIP_NO_TALENT = "No Talents"
 L_TOOLTIP_SPELL_ID = "Spell ID:"
--- TOTAL MEMORY LOCALIZATION
+-- Total Memory Localization
 L_TOTALMEMORY_USAGE = "Total Memory:"
--- TUTORIAL LOCALIZATION
+-- Tutorial Localization
 L_TUTORIAL_MESSAGE_1 = "For technical support visit https://github.com/Kkthnx."
 L_TUTORIAL_MESSAGE_2 = "You can toggle the microbar by using your right mouse button on the minimap."
 L_TUTORIAL_MESSAGE_3 = "You can set your keybinds quickly by typing /kb."
@@ -209,15 +211,15 @@ L_TUTORIAL_MESSAGE_7 = "To setup which channels appear in which chat frame, righ
 L_TUTORIAL_MESSAGE_8 = "You can use the /resetui command to reset all of your movers. You can also type /moveui and just right click a mover to reset its position."
 L_TUTORIAL_MESSAGE_9 = "To move abilities on the actionbars by default hold shift + drag. You can change the modifier key from the actionbar options menu."
 L_TUTORIAL_MESSAGE_10 = "You can see someones average item level of their gear by enabling the item level for tooltip option"
--- WOWHEAD LINK LOCALIZATION
+-- Wowhead Link Localization
 L_WATCH_WOWHEAD_LINK = "Wowhead Link"
--- WELCOME LOCALIZATION
+-- Welcome Localization
 L_WELCOME_LINE_1 = "Welcome to |cff71d5ffKkthnxUI|r v"
 L_WELCOME_LINE_2_1 = ""
 L_WELCOME_LINE_2_2 = "Type /cfg to config interface, or visit www.github.com/kkthnx|r"
 L_WELCOME_LINE_2_3 = ""
 L_WELCOME_LINE_2_4 = "Some of your questions can be answered by typing /uihelp"
--- SLASH COMMANDS LOCALIZATION
+-- Slash Commands Localization
 L_SLASHCMD_HELP = {
     "",
     "|cff71d5ffAvailable slash commands:|r",